Fixing interaction handler tests
