+++ lib/msal-browser/src/interaction_handler/SilentHandler.ts
@@ -27,7 +27,7 @@ export class SilentHandler extends InteractionHandler {
             this.authModule.logger.info("Navigate url is empty");
             throw BrowserAuthError.createEmptyNavigationUriError();
         }
-        const frameName = `msalTokenFrame${userRequestScopes}`;
+        const frameName = userRequestScopes ? `msalTokenFrame${userRequestScopes}` : `msalTokenFrame`;
         return this.loadFrameTimeout ? await this.loadFrame(requestUrl, frameName) : this.loadFrameSync(requestUrl, frameName);
     }
