+++ lib/msal-browser/test/interaction_handler/PopupHandler.spec.ts
@@ -138,53 +138,4 @@ describe("PopupHandler.ts Unit Tests", () => {
             expect(browserStorage.getItem(BrowserConstants.INTERACTION_STATUS_KEY)).to.be.eq(BrowserConstants.INTERACTION_IN_PROGRESS_VALUE);
         });
     });
-
-    describe("handleCodeResponse()", () => {
-
-        it("throws error if given location hash is empty", async () => {
-            await expect(popupHandler.handleCodeResponse("")).to.be.rejectedWith(BrowserAuthErrorMessage.hashEmptyError.desc);
-            await expect(popupHandler.handleCodeResponse("")).to.be.rejectedWith(BrowserAuthError);
-
-            await expect(popupHandler.handleCodeResponse(null)).to.be.rejectedWith(BrowserAuthErrorMessage.hashEmptyError.desc);
-            await expect(popupHandler.handleCodeResponse(null)).to.be.rejectedWith(BrowserAuthError);
-        });
-
-        it("successfully handles response", async () => {
-            const testCodeResponse: CodeResponse = {
-                code: "testAuthCode",
-                userRequestState: `${RANDOM_TEST_GUID}|testState`
-            };
-            const idTokenClaims = {
-                "ver": "2.0",
-                "iss": `${TEST_URIS.DEFAULT_INSTANCE}9188040d-6c67-4c5b-b112-36a304b66dad/v2.0`,
-                "sub": "AAAAAAAAAAAAAAAAAAAAAIkzqFVrSaSaFHy782bbtaQ",
-                "exp": "1536361411",
-                "name": "Abe Lincoln",
-                "preferred_username": "AbeLi@microsoft.com",
-                "oid": "00000000-0000-0000-66f3-3332eca7ea81",
-                "tid": "3338040d-6c67-4c5b-b112-36a304b66dad",
-                "nonce": "123523"
-            };
-
-            const testAccount = new Account(idTokenClaims.oid, TEST_DATA_CLIENT_INFO.TEST_HOME_ACCOUNT_ID, idTokenClaims, TEST_TOKENS.IDTOKEN_V2);
-            const testTokenResponse: TokenResponse = {
-                accessToken: TEST_TOKENS.ACCESS_TOKEN,
-                idToken: TEST_TOKENS.IDTOKEN_V2,
-                scopes: ["scope1", "scope2"],
-                refreshToken: TEST_TOKENS.REFRESH_TOKEN,
-                account: testAccount,
-                expiresOn: new Date(Date.now() + (TEST_TOKEN_LIFETIMES.DEFAULT_EXPIRES_IN * 1000)),
-                idTokenClaims: idTokenClaims,
-                tenantId: idTokenClaims.tid,
-                tokenType: "Bearer",
-                uniqueId: idTokenClaims.oid,
-                userRequestState: "testState"
-            };
-            sinon.stub(AuthorizationCodeModule.prototype, "handleFragmentResponse").returns(testCodeResponse);
-            sinon.stub(AuthorizationCodeModule.prototype, "acquireToken").resolves(testTokenResponse);
-            
-            const tokenResponse = await popupHandler.handleCodeResponse(TEST_HASHES.TEST_SUCCESS_CODE_HASH);
-            expect(tokenResponse).to.deep.eq(testTokenResponse);
-        });
-    });
 });
